my_site/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-filter==23.5
django-ratelimit==4.1.0
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.27.0
flake8==6.1.0
flake8-docstrings==1.7.0
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mccabe==0.7.0
Pillow==10.1.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
snowballstemmer==2.2.0
sqlparse==0.4.4
uritemplate==4.1.1
